requirements.txt
PyYAML
PyMuPDF
discord
requests